containerd

+++ scripts/installContainerd.sh
@@ -8,6 +8,15 @@ install_ubuntu() {
        apt-get remove -y containerd
        apt-get purge -y containerd
        apt -y autoremove
+       if [ -f /etc/modules-load.d/containerd.conf ];then
+          rm /etc/modules-load.d/containerd.conf
+       fi
+       if [ -f /etc/sysctl.d/99-kubernetes-cri.conf ];then
+          rm /etc/sysctl.d/99-kubernetes-cri.conf
+       fi
+       if [ -f /etc/containerd/config.toml ];then
+          rm /etc/containerd/config.toml
+       fi
     else
        echo "containerd is not installed.. continue to install"
     fi
@@ -50,16 +59,24 @@ install_centos() {
 
 configure_contd() {
 
+if [ -f /etc/modules-load.d/containerd.conf ];then
+echo "/etc/modules-load.d/containerd.conf exist, skipping.."
+else
 cat <<EOF | sudo tee /etc/modules-load.d/containerd.conf
 overlay
 br_netfilter
 EOF
+fi
 
+if [ -f /etc/sysctl.d/99-kubernetes-cri.conf ];then
+echo "/etc/sysctl.d/99-kubernetes-cri.conf exists, skipping.."
+else
 cat <<EOF | sudo tee /etc/sysctl.d/99-kubernetes-cri.conf
 net.bridge.bridge-nf-call-iptables = 1
 net.bridge.bridge-nf-call-ip6tables = 1
 net.ipv4.ip_forward = 1
 EOF
+fi
 
 sudo modprobe overlay
 sudo modprobe br_netfilter
@@ -67,11 +84,20 @@ sudo modprobe br_netfilter
 #reload kernel config 
 sudo sysctl --system
 # generate default containerd config 
-mkdir -p /etc/containerd
-containerd config default | sudo tee /etc/containerd/config.toml
-sed -i "s/SystemdCgroup = false/SystemdCgroup = true/g" /etc/containerd/config.toml
-systemctl restart containerd
-
+if [ -d /etc/containerd ];then
+   if [ -f /etc/containerd/config.toml ];then
+      echo "containerd default config exist, skipping.."
+   else
+      containerd config default | sudo tee /etc/containerd/config.toml
+      sed -i "s/SystemdCgroup = false/SystemdCgroup = true/g" /etc/containerd/config.toml
+      systemctl restart containerd
+   fi
+else
+   mkdir -p /etc/containerd
+   containerd config default | sudo tee /etc/containerd/config.toml
+   sed -i "s/SystemdCgroup = false/SystemdCgroup = true/g" /etc/containerd/config.toml
+   systemctl restart containerd
+fi
 }
 ################ MAIN ###################
